% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccade_recognition.R
\docType{data}
\name{samples}
\alias{samples}
\title{Samples of Eye Positions as Recorded with an Eye-Tracker}
\format{a data frame containing one line per sample.  The samples
are sorted in chronological order.  Time is given in milliseconds,
x- and y-coordinates in screen pixels.}
\source{
Recorded with an iViewX Eye-Tracker by SMI at approximately
250 Hz.
}
\usage{
samples
}
\description{
Samples of eye positions as recorded with an iViewX eye-tracker
recording at approx. 250 Hz.  The data quality is low on purpose
and contains episodes of track-loss and blinks.
}
\author{
Titus von der Malsburg \email{malsburg@posteo.de}
}
