% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saccade_recognition.R
\docType{package}
\name{saccades}
\alias{saccades}
\alias{saccades-package}
\title{Detection of Fixations in Raw Eye-Tracking Data}
\description{
Functions for the detection of fixations in raw eye-tracking data.
}
\details{
Offers a function for detecting fixations in a stream of eye
positions recorded by an eye-tracker.  The detection is done using
an algorithm for saccade detection proposed by Ralf Engbert and
Reinhold Kliegl (see reference below).  Anything that happens
between two saccades is considered to be a fixation.  This software
is therefore not suited for data sets with smooth-pursuit eye
movements.
}
\references{
Ralf Engbert, Reinhold Kliegl: Microsaccades uncover the
orientation of covert attention, Vision Research, 2003.
}
\seealso{
\code{\link{detect.fixations}},
\code{\link{diagnostic.plot}}, \code{\link{calculate.summary}}
}
\author{
Titus von der Malsburg \email{malsburg@posteo.de}
}
\keyword{classif}
\keyword{manip}
\keyword{ts}
